move metrics initialization
