+++ packages/service-core/package.json
@@ -18,9 +18,7 @@
   "dependencies": {
     "@js-sdsl/ordered-set": "^4.4.2",
     "@opentelemetry/api": "~1.8.0",
-    "@opentelemetry/exporter-metrics-otlp-http": "^0.51.1",
     "@opentelemetry/exporter-prometheus": "^0.51.1",
-    "@opentelemetry/resources": "^1.24.1",
     "@opentelemetry/sdk-metrics": "1.24.1",
     "@powersync/lib-services-framework": "workspace:*",
     "@powersync/service-jpgwire": "workspace:*",
