+++ service/package.json
@@ -11,9 +11,11 @@
   },
   "dependencies": {
     "@fastify/cors": "8.4.1",
-    "@opentelemetry/api": "~1.6.0",
-    "@opentelemetry/exporter-prometheus": "^0.43.0",
-    "@opentelemetry/sdk-metrics": "^1.17.0",
+    "@opentelemetry/api": "~1.8.0",
+    "@opentelemetry/exporter-prometheus": "^0.51.1",
+    "@opentelemetry/sdk-metrics": "1.24.1",
+    "@opentelemetry/exporter-metrics-otlp-http": "^0.51.1",
+    "@opentelemetry/resources": "^1.24.1",
     "@powersync/service-core": "workspace:*",
     "@powersync/lib-services-framework": "workspace:*",
     "@powersync/service-jpgwire": "workspace:*",
